Improve DisruptTripleoNodesTest tearDown
This tearDown is only executed when
test_z99_reboot_controller_galera_main_vip test is run.
It tries to delete the group of VMs created for this test, whose names
have a common pattern: "group_of_vms_%d" % i
Sometimes the vms_detailed_info variable does not contain all the
existing VM IDs and, due to this, some of the VMs are not deleted.
With this patch, whenever vms_detailed_info does not include the ID for
some VM, the tearDown method will search for it using its VM name.

Change-Id: Id7b2e4cc0d7b0e9deb7a8865f9c891de3f38260f