Move sphinx_rtd_theme to the readthedocs requirements
This package was removed from the OpenStack requirements with [1] so
keeping it in the doc/requirements.txt file is causing failure in the
requirements-check job.
As we want to use this theme to build our docs, lets keep it in the
readthedocs_requirements.txt file which isn't checked by the
requirements-check job at all.

This patch also changes docs and releasenotes configs in the way that it
will fallback to one of the default themes provided by sphinx.

Finally it makes some additional changes to the releasenotes conf.py
file to make build of the releasenotes working fine.

[1] https://review.opendev.org/c/openstack/requirements/+/893665

Change-Id: Iff03b8596a9e772940fd959efa7cbf80a68e9eab