Skip background ping checks during controllers reboot with nonDVR
When DVR is disabled, rebooting or applying network disruptions on the
controllers nodes can affect dataplane traffic because the packets are
not directly routed to the compute nodes.
Due to this, some faults tests fail during the background ping checks.
This patch skips the background ping checks on those tests.

Change-Id: If51016ce7c3562d1f18ac1f2124db72ef29c90d7