Skip SharedFixture.cleanUp during check-resources steps
The SharedFixture.cleanUp method is called when setup_fixture fails. The
aim of this is to delete the resources that were not successfully
created and allow next tests to create them properly.
This is not valid with check-resources steps because resources are not
created again.
So, in this case, it makes more sense to skip deletion of resources and
run the rest of tests with the existing ones, even if they fail.

Change-Id: I11b4eb5178c9891a692836ef720ca0487f407e60