Make ubuntu VM interface name configurable
Depending on the hypervisor/compute were the Ubuntu VM instances are
spawned, their interface names are created following differnt
conventions. Tobiko decides which is the expected name depending on the
OSP version, but that is not a valid method to decide it in some cases.
This patch allows to configure the expected Ubuntu interface name.

Change-Id: I6d876c73d883dc588360d6f6bc54f952d0fa9e84