Increase retry timeout to obtain multi_ip_test_fixture.vms_detailed_info
Test test_z99_reboot_controller_galera_main_vip creates a set of 10 VMs.
While the creation is in progress, the test reboots some controller
nodes. Then, it waits until all VMs are created. It uses a timeout
because sometimes not all the VMs reach the ACTIVE status due to the
controller rebooted.
The timeout value of 4 minutes seems to be not enough in some cases.
This patch increases that value to 5 minutes.

Change-Id: I40fa9798f71d82047237e914373016122f54a392