Add RHOSP topology config options
This patch adds "rhosp" config option group and moves some
overcloud related config options to that new group.
Those options can be used in both overcloud nodes in the Tripleo
based deployments as well as in the EDPM nodes in the "NextGen"
OSP deployments.

This patch deprecates old, "overcloud_* options names and "tripleo"
group for those options.

Additionally it also removes "podified" group as there are no config
options only for the "podified" deployments for now.
There is no need to deprecate this group or option which was there as
those were introduced just now and not released yet at all.

Related-Jira: #OSP-22166
Change-Id: I0587f85f6f5c86cf547c04c339b314720606fb74