Move network_undisrupt_controller_main_vip to tearDown
The faults test test_network_disruptor_main_vip runs a method to undo
the disruptive action that it executes.
That method is called at the end of the test. If something fails before,
the "undisruption" is not called.
This patch moves network_undisrupt_controller_main_vip to the tearDown
method, which guarantees it would be executed, regardless of passing or
failing.

Change-Id: I05183de74d12a00bc57efd8a3bce10525a3efd0b
(cherry picked from commit 11a625d2153fc131ac2de2679646b2b3e69b9fa3)