[Podified] Make Topology class to decide about container runtime
Previously it was only used in the Tripleo Topology class and containers
module decided about runtime (docker or podman) which should be used.
Now we have also Podified deployments and we can't rely on checks like
"has_overcloud()" to decide about it.

This patch moves some containers related things to the rhosp and/or
podified modules and let topology class to decide what containers
runtime should be used.

Related-Jira: #OSP-22166
Change-Id: I7187de8884e44eaefcdda6fae160e89242761f39