Use configured process timeout when closing a process
Some processes get blocked waiting while they cannot read from either
stdout or stderr. The command timeout was not applied when tobiko tried
to close those processes.
With this patch, those processes will be closed due to timeout and the
tobiko retries mechanism would be able to retry the command.

Change-Id: I1836de50cbbe029f7c474598ec57ea2e52822dd1