Set correct tox_min_version for the stable/2023.2 branch
After [1] we started using upper constraints from the OpenStack
requirements in the tobiko ansible roles. It cause issue with
installation of the tox >= 4.13 as it is not compatible with
requirements from upper-constraints for this branch.

This patch adds setting tox_min_version to the 4.5.1 which should be ok
with upper-constraints from the stable/2023.2 branch of the OpenStack
services.

[1] https://review.opendev.org/c/x/tobiko/+/912718

Related: #2058378
Change-Id: I215d7e3da4fff0fd9ae7bf829070b346c83fa5cd