Add Tobiko CLI tool to execute tobiko tools
This patch adds base `tobiko` CLI tool and one command for it which is
`ping`.
Using this new CLI tool it is now possible to do `tobiko ping <server>`
to run tobiko function to ping some server and store results in the
file. This can be used e.g. to run background ping process in the POD on
OpenShift and use that in the Tobiko tests.

Related: #TOBIKO-100
Change-Id: I33aea73cead5a4f406ce1a08c819fe056e17e9a1