Let topology to decide how to start background ping
Previously background ping was implemented only for the tripleo topology
and tests which were using it had to be skipped if were run on different
topology.
With this patch it is topology class who decides how background ping is
going to be started.
Currently it is only supported by the Tripleo topology and using
undercloud to run background ping but in the future it will be supported
also by the podified topology to run background ping in the POD.

Related: #TOBIKO-100
Change-Id: Ie60d3d066509c5579474dec482f039a7713a41a0