Add config option 'run_background_ping_in_pod'
This new config option can be used to force run background ping process
on the OCP POD, even if used Topology is Tripleo based.
For now running this background ping from the POD is not yet supported,
that will be added in next patches.

This option will be useful to run background ping from the OCP POD
during the adoption CI jobs so that results will be available for tobiko
run before adoption (tripleo topology) and after adoption (podified topology).

Related: #TOBIKO-100
Change-Id: I30f1be9b426af171f1925dc082bd5ff2853f6023