Cleanup of infrared zuul jobs
tobiko-infrared-ubuntu is removed. It always fails because infrared is
not compatible with ubuntu-focal.

Former tobiko-infrared is removed. It did not really test anything
because there was no "primary" host in the inventory and essential
ansible plays from this job need to be run on the "primary" host.

Former tobiko-infrared-centos-9 is renamed to tobiko-infrared and it
will only use one node. The aim of this job is just to validate the the
tobiko infrared plugin works (it just executes tobiko unit and
functional tests).

Existing tobiko-check-collected-files ansible role has been removed. It
was only used to test that results where properly generated when the
tobiko infrared plugin was used, but it only work when no workflow was
specify.

Change-Id: I6e334014420757aafa436f42b189c932ed02ae56