Cleanup tobiko jobs
Removed devstack-tobiko-minimal, as it was just the same as
devstack-tobiko, but on setups with lower requirements, and this caused
some inestabilities recently.

Removed devstack-tobiko-heat which tested functional tests
(devstack-tobiko-functional already covers this) and create-resources
tests, which are covered in many other jobs.

Removed devstack-tobiko-sanity, because the sanity tests are run by many
other jobs (those running the faults workflow, for example).

Removed devstack-tobiko-designate jobs because designate tests are
included into devstack-tobiko and devstack-tobiko-ovs.

Change-Id: Id8ee0323ad09297c08685ce19a60df469ae16826