Move rename_ping_staistics_file_to_checked function to common os module
This patch moves function `rename_ping_staistics_file_to_checked` to the
`tobiko.common._os` module and renames it to the `truncate_file`.
Additionally it adds helper function to get name of the truncated file.
The reason for those changes is that the same functions will be used in
the background iperf tests also.

Related: #TOBIKO-128
Change-Id: I14bbfc51f37323a106bd5e0e4fbf6d7145f716d8