Add AdvancedPeerServerStackFixture fixture
This new fixture creates 2 VMs using an "advanced" image, like e.g.
Fedora. One of the VMs has floating IP associated to it so and the other
one is only connected to the private network.
This new fixture will be needed for the network connectivity background
tests once e.g. iperf3 tests will be implemented there as iperf3 is not
available in the Cirros based VMs.

Related: #TOBIKO-128
Change-Id: I4b418ad71ac6853bdbf054e4047308b755b51ba6