Add East-West traffic background iperf3 test
This patch adds east-west traffic test using iperf3 client and server
running in background. When test will be run first time, it will simply
make sure that iperf3 client and server are run in the VMs. In the
following runs it will check if iperf3 client is running, if yes, stops
it, checks log file and starts it again.

Closes: #TOBIKO-129
Change-Id: I2f51e28f5fd06bceb845f1bf4bdff79ceaa3767a