Add Dockerfile for the iperf3 container image
When iperf3 background process will be run in the POD by Tobiko tests,
it needs to be newer iperf3 then available in Centos 9 stream currently
(Tobiko requires iperf3 with support for the "--json-stream" option
which was introduced in iperf 3.17).
Because of that this patch adds simply Dockerfile to build iperf3
container. It is based on Centos 10 stream and has iperf3 in the new
enough version.

Related: #TOBIKO-131
Change-Id: Ie7b67bad059d7d5252270bab3cf148103dff96e9