Add support for the "--json-stream" option in iperf3 parameters
This option can be set in the command's parameters instead of the
"--logfile <logfile>" if special string "json-stream" will be provided
as parameter for the Iperf3Parameters class.
This new option is supported by iperf3 since iperf3 3.17 and is required
for the iperf3 run in the POD which will be introduced soon.

Related: #TOBIKO-131
Change-Id: I828aa6d0468e81d26244ff4437141baa211e2c07