Implement way to run iperf3 client in the POD
This patch implements functions necessary to start and stop iperf3
client in the POD on OpenShift. It also implements function to convert
logs from the iperf3 POD's stdout (json-stream), save them locally and
then check this log in the same ways as it is done for the iperf3
running on the guest VM or locally on the server where Tobiko runs.

This patch also adds new config option "iperf3_image" which can be used
to specify container image to run iperf3 client. By default it is the
Tobiko image which have installed iperf3. But currently Tobiko image is
based on Centos9 stream and have too old version of iperf3 (3.17 is
requried at least as this version provides "--json-stream" option).
Because of that it is recommended to use different image for now, like
e.g. [1] which is not official but works fine.

[1] https://quay.io/repository/skaplons/iperf3

Related: #TOBIKO-131
Change-Id: I440c9438954ea8e9e6159d7a1df472539e471988