_check_ping_results waits until log file is available
At the end of check_or_start_tobiko_command, the background ping pod is
started and immediately it checks whether the ping log file is available.
This may fail due to a race condition: the pod may take a few seconds to
start and the ping log file may be unavailable for a little more time.

This patch adds retries until the ping log file is available.

Change-Id: I9b882082b0506e8b0c6389eea8199f644bb65df2