Retry to obtain the iperf log file until it is not empty
This applies to the east-west iperf test, where iperf runs from a VM
instance and it writes results to a file when the process is killed, but
it may take some seconds to write the results to the file after being
killed.

Besides, with this patch the results file is first downloaded to the
local test machine, in order to avoid issues with huge files.

Change-Id: Ie68f88d7da2370d7faafe4ef072d7cedac695f22