Run background iperf3 tests with --json-stream option always
Either iperf3 is run from a pod (we recommend to use a certain centos10
pod) or from the supported Fedora 41 Advanced VM instances, iperf3
supports the --json-stream option on those environments (iperf version
greater or equal to 3.17).
This option can be used with --logfile (the output is written for every
iperf3 interval on a file) or without it (the output is written to
stdout).
This option should not be used when iperf3 is run from the test
machine/container: undercloud, devstack, tobiko pod (in case of podified
envs).

Change-Id: I0491dbff33619ca3096a9eeac70ae4a46c68b62d