Create VLAN on different parent interfaces for Customize Fedora image
Fedora VM instance's interface name is different depending on the
compute OS. Tobiko customizes a Fedora image to create a vlan101 on top
of its NIC when the VM boots.
With this patch, both potential NIC names (eth0 and ens3) are considered.

nmcli will create two connections with same name and different UUIDs and
only one of them will become active: the one that finds an existing
parent NIC.

Change-Id: I3eac916579c7add0a2f397e796e8548028d81aab