Add "http ping" command to the Tobiko CLI tool
This patch adds new CLI command `tobiko http ping <url>` which will be
used in the background tests. It is simply script which sends HTTP
request to the given url every N seconds given as an `interval`
parameter. It only do HTTP HEAD request to check response code. If the
response code is 2xx or 3xx then this tool considers it as OK. If it is
4xx or higher, then it considers it as FAILED.

Results are stored in the file which can be also configured through the
CLI option.

Related: #TOBIKO-130
Change-Id: I7f40c1fb068856e6b8cc04f8d5ae41c9ebe3a8e6