Implement http_ping background tests run from the POD
This patch implements way to run background HTTP ping test inside the POD.
It is implemented for the podified and tripleo topologies.
It is not supported for the devstack topology.

This patch adds also test to run http pings in background: for the North-South
traffic - this tests if new TCP connections from the external network to the VM
can be established using Floating IP,

Closes: #TOBIKO-130
Closes: #TOBIKO-133
Change-Id: I13b1a92669d2e4553cd395a2ed30bd89deb7c909