Use maxprocesses instead of numprocesses pytest-xdist option
Until now, we could set pytest --numprocesses option to any value
different from "auto", by setting TOX_NUM_PROCESSES to an integer value.
This was problematic when that value was higher than the actual number
of available CPUs.

With this patch, the TOX_NUM_PROCESSES env variable is used to set
--maxprocesses, and --numprocesses is always set to "auto". Hence, the
actual number of tobiko processes will be the lower value between "the
number of available CPUs" and TOX_NUM_PROCESSES.

Change-Id: Ic4dc692fb0a0b7c6177208630609af625de37435