Add 'interval' option to the iperf3 run in background
With 'interval' parameter passed to the iperf3 client we can control
how often it will report results to the stdout or file. By default it is
1 second which is good in typical cases but not that good for our
background tests which can run for very long time and such log can be
really huge.
With this patch this can be configured in Tobiko. By default it is still
set to the default iperf3 value (1s) but can be customized through the
Tobiko config.

Related: #TOBIKO-147
Change-Id: I6fc984a66ba839aebacbd0e0929876a17166b619