Add 'interval' option to the tobiko ping CLI tool
With this new option it is possible to specify interval for the ping
command run in background (e.g. in the POD). It is interval between
two consequentive ICMP packtes. Making this interval longer then default
1 second will decrease size of the log file of the background ping
process when it is run e.g. during the adoption CI job.

Value of this parameter can be configured in Tobiko config file using
tobiko.config.ping.interval option which already existed in Tobiko
before this patch. This patch only added possibility to pass it to the
CLI tool `tobiko ping` as that tool typically don't use same config file
as Tobiko test runner.

Related: #TOBIKO-147
Change-Id: If8f1880a1ee53739dfce5a9e51c90db33403c46d