Fix extra-requirements.txt for pandas versions
pandas was already pinned to 2.1.1 for python>=3.9, but it also tried to
download and build wheel the latest pandas version (although later it
was not installed) due to the previous line from
extra-requirements.txt.

That previous line has been modified to affect only python<3.9 and to
install a specific version of pandas, 2.0.3.

Change-Id: I6692182e52a852368dd491e67a35efb729b5fe85