Reduce retry interval from 5 to 3 for all loops involving OSP API requests
Apparently, sending requests to Openstack APIs on Podified environments
every 5 seconds is problematic and tobiko tests are unstable due to
this.
In two previous patches, this change was already done for heat and
glance requests.
With this patch, the interval has been changed for all loops that
involve retries to any Openstack APIs.

OSPCIX-844

Change-Id: If0bb72b6e46683425429e3b93d3993fb57221799