Configure tobiko clients to close TCP connections
This patch adds a header that forces openstack component clients used by
tobiko to close TCP connection when each one of their requests is
answered. Reusing TCP connections has been problematic on podified
environments and we will try to avoid it.

It has been applies to glance, heat, ironic, neutron and nova. This
option was not available for manila and keystone.

Related-ticket: OSPCIX-844
Change-Id: I3e496a3efc355332945454eefaee4d9f202b9d33