Kill all dhcp_ping scripts found on the VM
Previously Tobiko was looking only for one PID with the dhcp ping script
running on the guest VM and was killing that process when it was needed.
Apparently there may be also fork of the original process visible in the
"ps" output sometimes and if that is still alive when Tobiko is trying
to start dhcp_ping script again, it wasn't starting script at all, then
forked process was also gone and Tobiko failed with the test as
dhcp_ping script wasn't running as expected.

Now this should be fixed as Tobiko is looking for all dhcp_ping
processes and tries to kill all of them so no fork should be visible
during the star of the new script.

Closes: #TOBIKO-177
Change-Id: I7620172bc792c80f3b288811940f9f620e9ee719