Better obtain MTU from neutron rather than from heat
Tobiko obtains MTU from heat stacks that it uses to create networks.
That is ok unless the network MTU is modified.
Running OVS to OVN migration modifies the existing network MTUs
in some cases. Example: when a vxlan network becomes a geneve network
(the MTU should be lowered by 8 because of the geneve headers length).

With this patch, tobiko will obtain the MTU from neutron.
In cases like OVN migration, the tests will use the updated MTU value
and will pass.

Change-Id: Id9e2b9a2852e0e145c98e54122888194d2235e15