Add retries when checking if background iperf3 server started
Before this patch tobiko sent the background iperf3 server command and
checked immediately whether the process was running. This could
sometimes fail, depending on the VM load.
With this patch, tobiko retries 5 times to check if the process
started.

A similar patch applied on background ping was recently created too:
https://review.opendev.org/c/x/tobiko/+/964519

Closes: #TOBIKO-188

Change-Id: I530c44ce933c7621500cb06eef8d6c10259b582e
Signed-off-by: Eduardo Olivares <eolivare@redhat.com>